// client/src/types/common/IQuotationSummary.ts
export default interface IQuotationSummary {
	fundType: string | number | readonly string[] | undefined;
	documentNo: string;
	date: string;
	//   auto generated
	epfNo: string;
	division: string;
	quotationRequestNo: string;
	fileNo: string;
	srnNo: string;
	value: string;
	fund: string;
	project: string;
	remark: string;
}